Removes current IsDirectory implementation to only handle FilesPaths DirectoryPath will be included in a separate implementation
