+++ Xamarin.PropertyEditing/Common/FilePath.cs
@@ -1,11 +1,10 @@
 ﻿using System;
 
-namespace Xamarin.PropertyEditing
+namespace Xamarin.PropertyEditing.Common
 {
-	public class FilePath
+	public class FilePath : IEquatable<FilePath>
 	{
 		public string Source { get; set; }
-		public bool IsDirectory { get; set; }
 
 		public FilePath () { }
 
@@ -17,6 +16,27 @@ namespace Xamarin.PropertyEditing
 		public override string ToString ()
 		{
 			return Source;
+		}
+
+		public override bool Equals (object obj)
+		{
+			if (obj == null) return false;
+			if (!(obj is FilePath)) return false;
+			return Equals ((FilePath)obj);
+		}
+
+		public bool Equals (FilePath other)
+		{
+			return Source.Equals (other.Source, StringComparison.InvariantCultureIgnoreCase);
+		}
+
+		public override int GetHashCode ()
+		{
+			var hashCode = 1861433795;
+			unchecked {
+				hashCode = 1861433795 * -1521134295 + Source.GetHashCode ();
+			}
+			return hashCode;
 		}
 	}
 }
